massive cleanups debian/rules
For this upstream release of ffmpeg, various cleanups are
needed. Moreover, since this branch now follows the 'twin-branch'
approach (master and master.unstripped), a lot of switches and cases had
to be rewritten.

Documentation is now always built, since we package svn checkouts and no
longer release tarballs. Moreover building documentation now requires
the tree to be configured before the  documentation can be built.

The faad configure flag was renamed.