remove debconf question mplayer/dvd_device
Modern Debian systems are almost always managed using udev, which
provides a /dev/dvd symlink. Relying on this is certainly better than
bugging the user about this.

For the case that the user has more than one dvd drive and udev gets the
wrong one, he can still override that decison in his configuration file.

For the case that the system doesn't use udev, we can assume that the
user is experienced enough to manage the configuration file himself.