add replaces field for mplayer-doc and mplayer
Since the translated manpages are now no longer in the mplayer package
itself, but moved in the mplayer-doc package, we need to add a proper
Replaces field.

Also, most files formerly shipped in the mplayer package has been moved
now to mplayer-nogui, it must also declare a proper replaces field on
the packages mplayer and mplayer-doc.