Revert "use upstream's Makefile to install the manpage"
Upstream installs the manpage as symlink. Because in debian we compress
them, the name changes and the symlink points to the void then. The
previous aproach avoids this by placing a copy of the manpage
there. This is technically still a bit of a waste, but at least leaves
us with a working manpage.

Error found by lintian.

This reverts commit 676627237477b5725b7c0f7808c0a3c36df0f6e7.