remove DEB_BUILD_CONFIGURE override hack
This is done to simplify handling of DEB_BUILD_CONFIGURE and
greatly simplify handling of debian/README.Debian

This file used to be generated depending on the name of the source
package. This patch installs a static copy of README.Debian with only
relevant comments merged from the other fragments.